requirements.txt
# ==================== Core AI/LLM ====================
langchain==0.2.16
langchain-core==0.2.38
langchain-groq==0.1.9
langchain-community==0.2.16
groq>=0.4.0,<1.0.0  # ✅ ADD: Upper bound for safety

# ==================== Web Framework ====================
streamlit==1.39.0

# ==================== Database ====================
psycopg2-binary>=2.9.9,<3.0.0  # ✅ ADD: Upper bound

# ==================== Utilities ====================
python-dotenv==1.0.1
requests==2.32.3

# ==================== PDF Generation ====================
reportlab==4.0.7

# ==================== Build Tools ====================
setuptools==75.2.0

# ==================== Optional: Development Tools ====================
# pytest==8.0.0  # For testing
# black==24.0.0   # Code formatting
# flake8==7.0.0   # Linting
